% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcaCor.R
\name{fpcaCor}
\alias{fpcaCor}
\title{Functional Principal Analysis using Correlation Matrix}
\usage{
fpcaCor(
  X = NULL,
  types = "con",
  argvals = NULL,
  nbasis = 10,
  pve = 0.99,
  npc = NULL
)
}
\arguments{
\item{X}{a numeric matrix (n x p). It is supplied by user. Here n is the number of samples and p is the number of observations.}

\item{types}{a vector of length p representing the type of each of the p variables in \code{X}. For our case, it is "con". Users can learn more about it here: \code{\link[latentcor]{latentcor}}.}

\item{argvals}{the argument values of the function evaluations in Y, defaults to a equidistant grid from 0 to 1.}

\item{nbasis}{number of B-spline basis functions used for estimation of the mean function and bivariate smoothing of the covariance surface.}

\item{pve}{proportion of variance explained: used to choose the number of principal components. It should be supplied by users.The default is 0.99. (See \code{\link[refund]{fpca.sc}}.)}

\item{npc}{the number of principal components.if it is given, this overrides pve; the default is NULL. (See \code{\link[refund]{fpca.sc}}.)}
}
\value{
A list with the elements
\item{npc}{The number of principal components. If it is given, the given value. Otherwise it is calculated in the function based on given \code{pve}}
\item{eigenfuncs}{A (p x npc) matrix where its colums are eigenfunction extracted from the smoothed correlation matrix.}
}
\description{
Extract eigen-functions from smoothed correlation matrix which comes from a Gaussian latent model.
}
\examples{
#To see the difference when we have 2 different pve values.
#Create a matrix
X = matrix(rnorm(1000), ncol = 25)
#Example 1
fpcaCor(X = X, types = "con", argvals = NULL, nbasis = 10, pve = 0.99, npc = NULL)

#Example 2
fpcaCor(X = X, types = "con", argvals = NULL, nbasis = 10, pve = 0.95, npc = NULL)

}
