% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_copula_cor.R
\name{gaussian_copula_cor}
\alias{gaussian_copula_cor}
\title{Pointwise Correlation from Gaussian latent model}
\usage{
gaussian_copula_cor(n = 100, ntime = 50)
}
\arguments{
\item{n}{A positive integer indicating the sample size. The default value is 100.}

\item{ntime}{A positive integer indicating the number of time points. The default value is 50.}
}
\value{
A list with the elements
\item{Y}{A n x ntime matrix for the Gaussian latent model
\item{}{A ntime x ntime pointwise correlation matrix for the Gaussian latent model}
}
\description{
Pointwise Correlation from Gaussian latent model
}
\examples{
gaussian_copula_cor(n = 100, ntime = 20)
gaussian_copula_cor(n = 20, ntime = 2)
}
