% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_quadWeights.R
\name{my_quadWeights}
\alias{my_quadWeights}
\title{Quadrature weights}
\usage{
my_quadWeights(argvals, method = "trapezoidal")
}
\arguments{
\item{argvals}{function arguments}

\item{method}{quadrature method. Can be either trapedoidal or midpoint.}
}
\value{
a vector of quadrature weights for the points supplied in argvals.
}
\description{
Necessary function for numerical integration.
}
