% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_spline_mean.R
\name{smooth_spline_mean}
\alias{smooth_spline_mean}
\title{Smooth spline mean}
\usage{
smooth_spline_mean(Y = NULL, argvals = NULL, center = TRUE)
}
\arguments{
\item{Y}{a numerical matrix which is given by users.}

\item{argvals}{a numerical vector which can be given by users.}

\item{center}{a logical argument, it is TRUE (default).}
}
\value{
smoothed mean of given matrix Y.
}
\description{
Smooth spline mean
}
\examples{
### settings
n = 50  # number of subjects
p = 200  #number of time points
Y = matrix(rnorm(n*p), nrow = n, ncol = p)
t = (1:p)/p  # a regular grid on [0,1]
result = smooth_spline_mean(Y = Y, argvals = t)
}
