% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{soft threshold function}
\usage{
soft(a, lambda)
}
\arguments{
\item{a}{A scalar to be thresholded}

\item{lambda}{Non-negative parameter}
}
\value{
Returns S(a, lambda)
}
\description{
soft threshold function
}
\examples{
soft(3,2)
soft(1,2)
}
