% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.cor.R
\name{fpca.cor}
\alias{fpca.cor}
\title{Functional Principal Analysis using Correlation Matrix}
\usage{
fpca.cor(
  X = NULL,
  types = "con",
  argvals = NULL,
  nbasis = 10,
  pve = 0.99,
  npc = NULL
)
}
\arguments{
\item{X}{a numeric matrix(I x D). It is supplied by user. Here I is the number of samples and D is the number of observations.}

\item{types}{a vector of length D representing the type of each of the D variables in \code{X}.For our case, it is "con". Users can learn about it here: \code{\link[latentcor]{latentcor}}.}

\item{argvals}{the argument values of the function evaluations in Y, defaults to a equidistant grid from 0 to 1.}

\item{nbasis}{number of B-spline basis functions used for estimation of the mean function and bivariate smoothing of the covariance surface.}

\item{pve}{proportion of variance explained: used to choose the number of principal components. It should be supplied by users.The default is 0.99.}

\item{npc}{the number of principal components.if it is given, this overrides pve; the default is NULL.}
}
\value{
eigen-functions A list of vectors (eigen-vectors) of length I extracted from the smoothed correlation matrix.
}
\description{
Extract eigen-functions from smoothed correlation matrix which comes from a Gaussian latent model.
}
\examples{
#Create a matrix
X = matrix(rnorm(1000, ncol = 25))
fpca.cor(X = X, types = "con", argvals = NULL, nbasis = 10, pve = 0.99, npc = NULL)

}
