% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_copula_cor.R
\name{gaussian_copula_cor}
\alias{gaussian_copula_cor}
\title{Pointwise Correlation from Gaussian latent model}
\usage{
gaussian_copula_cor(n = 100, ntime = 50)
}
\value{
pointwise correlation matrix
}
\description{
Pointwise Correlation from Gaussian latent model
}
\examples{
gaussian_copula_cor(n = 100, ntime = 20)
gaussian_copula_cor(n = 20, ntime = 2)
}
