% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_latentcor.R
\name{smooth_latentcor}
\alias{smooth_latentcor}
\title{Smoothed correlation matrix}
\usage{
smooth_latentcor(
  Y = NULL,
  argvals = NULL,
  Khat = NULL,
  cov.est.method = 2,
  useSymm = TRUE,
  nbasis = 10
)
}
\arguments{
\item{useSymm}{}
}
\value{

}
\description{
Smoothed correlation matrix
}
